pyproject.toml
[build-system]
requires = ["setuptools>=69.5", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taboo-gwen3"
version = "0.1.0"
description = "Replication framework for taboo secret suppression experiments on Qwen3 models."
readme = "README.md"
authors = [
  { name = "Luis Miguel Montoya" }
]
requires-python = ">=3.11"
dependencies = [
  "transformers>=4.41.2",
  "accelerate==0.31.0",
  "bitsandbytes==0.43.1",
  "trl==0.9.6",
  "peft==0.11.1",
  "datasets==2.20.0",
  "evaluate==0.4.2",
  "sentencepiece==0.2.0",
  "safetensors==0.4.3",
  "numpy==1.26.4",
  "pandas==2.1.4",
  "torch==2.3.1",
  "wandb==0.16.6",
  "mlflow==2.14.1",
  "pyyaml==6.0.1",
  "rich==13.7.1",
  "hydra-core==1.3.2",
  "tyro==0.7.3",
  "einops==0.8.0",
  "scikit-learn==1.4.2",
  "jsonlines==4.0.0"
]

[project.optional-dependencies]
deploy = [
  "fastapi==0.111.0",
  "uvicorn==0.30.1"
]
notebooks = [
  "jupyterlab==4.2.3",
  "notebook==7.2.1"
]
quality = [
  "black==24.4.2",
  "isort==5.13.2",
  "ruff==0.5.4",
  "mypy==1.10.0",
  "pytest==8.2.2",
  "pytest-cov==5.0.0",
  "types-PyYAML==6.0.12.20240311",
  "types-protobuf==4.24.0.20240311"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E203", "E266", "E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=taboo_gwen3 --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["taboo_gwen3"]
